Refactor eval Pandas and Spark df evaluation
