+++ lale/helpers.py
@@ -1111,127 +1111,3 @@ def _is_pandas_df(df):
 def _is_spark_df(df):
     if spark_installed:
         return isinstance(df, spark_df)
-
-class _PandasEvaluator(ast.NodeVisitor):
-    def __init__(self, X):
-        self.result = None
-        self.df = X
-
-    def visit_Constant(self, node: ast.Constant):
-        self.result = node.value
-
-    def visit_Subscript(self, node: ast.Subscript):
-        if _is_ast_name (node.value) and node.value.id == "it":
-            self.visit(node.slice)
-            column_name = self.result
-            if column_name is None or not column_name.strip():
-                raise ValueError("Name of the column cannot be None or empty.")
-            self.result = self.df[column_name]
-        else:
-            raise ValueError(
-                f"Unimplemented expression"
-            )
-
-    def visit_Attribute(self, node: ast.Attribute):
-        if _is_ast_name (node.value) and node.value.id == "it":
-            self.result = self.df[node.attr]
-        else:
-            raise ValueError(
-                f"Unimplemented expression"
-            )
-
-    def visit_BinOp(self, node: ast.BinOp):
-        self.visit(node.left)
-        v1 = self.result
-        self.visit(node.right)
-        v2 = self.result
-        if isinstance(node.op, ast.Add):
-            self.result = v1 + v2
-        elif isinstance(node.op, ast.Sub):
-            self.result = v1 - v2
-        elif isinstance(node.op, ast.Mult):
-            self.result = v1 * v2
-        elif isinstance(node.op, ast.Div):
-            self.result = v1 / v2
-        elif isinstance(node.op, ast.FloorDiv):
-            self.result = v1 // v2
-        elif isinstance(node.op, ast.Mod):
-            self.result = v1 % v2
-        elif isinstance(node.op, ast.Pow):
-            self.result = v1 ** v2
-        else:
-            raise ValueError(
-                f"""Unimplemented operator {ast.dump(node.op)}"""
-            )
-
-    def visit_Call(self, node: ast.Call):
-        functions_module = importlib.import_module("lale.lib.lale.functions_pandas")
-        function_name = node.func.id
-        map_func_to_be_called = getattr(functions_module, function_name)
-        self.result = map_func_to_be_called(self.df, node)
-
-def eval_pandas_df(X, expr):
-    evaluator = _PandasEvaluator(X)
-    evaluator.visit(expr._expr)
-    return evaluator.result
-
-class _SparkEvaluator(ast.NodeVisitor):
-    def __init__(self):
-        self.result = None
-
-    def visit_Constant(self, node: ast.Constant):
-        self.result = spark_lit(node.value)
-
-    def visit_Subscript(self, node: ast.Subscript):
-        if _is_ast_name (node.value) and node.value.id == "it":
-            column_name = node.slice.value.s  # type: ignore
-            if column_name is None or not column_name.strip():
-                raise ValueError("Name of the column cannot be None or empty.")
-            self.result = spark_col(column_name)
-        else:
-            raise ValueError(
-                f"Unimplemented expression"
-            )
-
-    def visit_Attribute(self, node: ast.Attribute):
-        if _is_ast_name (node.value) and node.value.id == "it":
-            self.result = spark_col(node.attr)
-        else:
-            raise ValueError(
-                f"Unimplemented expression"
-            )
-
-    def visit_BinOp(self, node: ast.BinOp):
-        self.visit(node.left)
-        v1 = self.result
-        self.visit(node.right)
-        v2 = self.result
-        if isinstance(node.op, ast.Add):
-            self.result = v1 + v2
-        elif isinstance(node.op, ast.Sub):
-            self.result = v1 - v2
-        elif isinstance(node.op, ast.Mult):
-            self.result = v1 * v2
-        elif isinstance(node.op, ast.Div):
-            self.result = v1 / v2
-        # elif isinstance(node.op, ast.FloorDiv):
-        #     self.result = v1 // v2
-        elif isinstance(node.op, ast.Mod):
-            self.result = v1 % v2
-        elif isinstance(node.op, ast.Pow):
-            self.result = v1 ** v2
-        else:
-            raise ValueError(
-                f"""Unimplemented operator {ast.dump(node.op)}"""
-            )
-
-    def visit_Call(self, node: ast.Call):
-        functions_module = importlib.import_module("lale.lib.lale.functions_spark")
-        function_name = node.func.id
-        map_func_to_be_called = getattr(functions_module, function_name)
-        self.result = map_func_to_be_called(node)
-
-def eval_spark_df(X, expr):
-    evaluator = _SparkEvaluator()
-    evaluator.visit(expr._expr)
-    return evaluator.result
