+++ lale/lib/lale/functions.py
@@ -134,211 +134,6 @@ class date_time:
         return result
 
 
-def replace(df: Any, replace_expr: Expr, new_column_name: str):
-    re: Any = replace_expr._expr
-    column_name = re.args[0].attr
-    if new_column_name is None:
-        new_column_name = column_name
-    mapping_dict = ast.literal_eval(re.args[1].value)
-    if _is_pandas_df(df):
-        new_column = df[column_name].replace(mapping_dict)
-        df[new_column_name] = new_column
-        if new_column_name != column_name:
-            del df[column_name]
-    elif spark_installed and _is_spark_df(df):
-        mapping_expr = create_map([lit(x) for x in chain(*mapping_dict.items())])  # type: ignore
-        df = df.withColumn(new_column_name, mapping_expr[df[column_name]])  # type: ignore
-        if new_column_name != column_name:
-            df = df.drop(column_name)
-    else:
-        raise ValueError(
-            "function replace supports only Pandas dataframes or spark dataframes."
-        )
-    return new_column_name, df
-
-
-def identity(df: Any, column: Expr, new_column_name: str):
-    if _is_ast_subscript(column._expr):  # type: ignore
-        column_name = column._expr.slice.value.s  # type: ignore
-    elif _is_ast_attribute(column._expr):  # type: ignore
-        column_name = column._expr.attr  # type: ignore
-    else:
-        raise ValueError(
-            "Expression type not supported. Formats supported: it.column_name or it['column_name']."
-        )
-
-    if column_name is None or not column_name.strip():
-        raise ValueError("Name of the column to be renamed cannot be None or empty.")
-    if new_column_name is None or not new_column_name.strip():
-        raise ValueError(
-            "New name of the column to be renamed cannot be None or empty."
-        )
-
-    if _is_pandas_df(df):
-        df = df.rename(columns={column_name: new_column_name})
-    elif spark_installed and _is_spark_df(df):
-        df = df.withColumnRenamed(column_name, new_column_name)
-    else:
-        raise ValueError(
-            "Function identity supports only Pandas dataframes or spark dataframes."
-        )
-    return new_column_name, df
-
-
-def ratio(df: Any, replace_expr: Expr, new_column_name: str):
-    def get_column_name(column):
-        if _is_ast_subscript(column):
-            column_name = column.slice.value.s
-        elif _is_ast_attribute(column):
-            column_name = column.attr
-        else:
-            raise ValueError(
-                "Expression type not supported. Formats supported: it.column_name or it['column_name']."
-            )
-        if column_name is None or not column_name.strip():
-            raise ValueError("Name of the column cannot be None or empty.")
-        return column_name
-
-    numerator_col = replace_expr._expr.args[0]  # type: ignore
-    denominator_col = replace_expr._expr.args[1]  # type: ignore
-    numerator_col_name = get_column_name(numerator_col)
-    denominator_col_name = get_column_name(denominator_col)
-
-    if new_column_name is None or not new_column_name.strip():
-        raise ValueError(
-            """New name of the column to be renamed cannot be None or empty. You may want to use a dictionary
-            to specify the new column name as the key, and the expression as the value."""
-        )
-
-    if _is_pandas_df(df):
-        df[new_column_name] = df[numerator_col_name] / df[denominator_col_name]
-    elif spark_installed and _is_spark_df(df):
-        df = df.withColumn(
-            new_column_name, df[numerator_col_name] / df[denominator_col_name]
-        )
-    else:
-        raise ValueError(
-            "Function identity supports only Pandas dataframes or spark dataframes."
-        )
-    return new_column_name, df
-
-
-def subtract(df: Any, replace_expr: Expr, new_column_name: str):
-    def get_column_name(column):
-        if _is_ast_subscript(column):
-            column_name = column.slice.value.s
-        elif _is_ast_attribute(column):
-            column_name = column.attr
-        else:
-            raise ValueError(
-                "Expression type not supported. Formats supported: it.column_name or it['column_name']."
-            )
-        if column_name is None or not column_name.strip():
-            raise ValueError("Name of the column cannot be None or empty.")
-        return column_name
-
-    op1_col = replace_expr._expr.args[0]  # type: ignore
-    op2_col = replace_expr._expr.args[1]  # type: ignore
-    op1_col_name = get_column_name(op1_col)
-    op2_col_name = get_column_name(op2_col)
-
-    if new_column_name is None or not new_column_name.strip():
-        raise ValueError(
-            """New name of the column to be renamed cannot be None or empty. You may want to use a dictionary
-            to specify the new column name as the key, and the expression as the value."""
-        )
-
-    if _is_pandas_df(df):
-        df[new_column_name] = df[op1_col_name] - df[op2_col_name]
-    elif spark_installed and _is_spark_df(df):
-        df = df.withColumn(new_column_name, df[op1_col_name] - df[op2_col_name])
-    else:
-        raise ValueError(
-            "Function identity supports only Pandas dataframes or spark dataframes."
-        )
-    return new_column_name, df
-
-
-def time_functions(
-    df: Any, dom_expr: Expr, new_column_name: str, pandas_func: str, spark_func: str
-):
-    fmt = None
-    de: Any = dom_expr._expr
-    column_name = de.args[0].attr
-    if new_column_name is None:
-        new_column_name = column_name
-    if len(de.args) > 1:
-        fmt = ast.literal_eval(de.args[1])
-    if _is_pandas_df(df):
-        new_column = pd.to_datetime(df[column_name], format=fmt)
-        df[new_column_name] = getattr(getattr(new_column, "dt"), pandas_func)
-        if new_column_name != column_name:
-            del df[column_name]
-    elif spark_installed and _is_spark_df(df):
-        df = df.withColumn(column_name, to_timestamp(df[column_name], fmt))  # type: ignore
-        df = df.select(eval(spark_func + "(df[column_name])").alias(new_column_name))
-        if new_column_name != column_name:
-            df = df.drop(column_name)
-    else:
-        raise ValueError(
-            "function day_of_month supports only Pandas dataframes or spark dataframes."
-        )
-
-    return new_column_name, df
-
-
-def day_of_month(df: Any, dom_expr: Expr, new_column_name: str):
-    return time_functions(df, dom_expr, new_column_name, "day", "dayofmonth")
-
-
-def day_of_week(df: Any, dom_expr: Expr, new_column_name: str):
-    return time_functions(df, dom_expr, new_column_name, "weekday", "dayofweek")
-
-
-def day_of_year(df: Any, dom_expr: Expr, new_column_name: str):
-    return time_functions(df, dom_expr, new_column_name, "dayofyear", "dayofyear")
-
-
-def hour(df: Any, dom_expr: Expr, new_column_name: str):
-    return time_functions(df, dom_expr, new_column_name, "hour", "spark_hour")
-
-
-def minute(df: Any, dom_expr: Expr, new_column_name: str):
-    return time_functions(df, dom_expr, new_column_name, "minute", "spark_minute")
-
-
-def month(df: Any, dom_expr: Expr, new_column_name: str):
-    return time_functions(df, dom_expr, new_column_name, "month", "spark_month")
-
-
-def string_indexer(df: pd.DataFrame, dom_expr: Expr, new_column_name: str):
-    de: Any = dom_expr._expr
-    column_name = de.args[0].attr
-    if new_column_name is None:
-        new_column_name = column_name
-
-    if _is_pandas_df(df):
-        sorted_indices = df[column_name].value_counts().index
-        df[new_column_name] = df[column_name].map(
-            dict(zip(sorted_indices, range(0, len(sorted_indices))))
-        )
-        if new_column_name != column_name:
-            del df[column_name]
-    elif spark_installed and _is_spark_df(df):
-        df = df.withColumnRenamed(
-            column_name, "newColName"
-        )  # renaming because inputCol and outputCol can't be the same.
-        indexer = StringIndexer(inputCol="newColName", outputCol=new_column_name)
-        df = indexer.fit(df).transform(df)
-        df = df.drop("newColName")
-    else:
-        raise ValueError(
-            "function day_of_month supports only Pandas dataframes or spark dataframes."
-        )
-
-    return new_column_name, df
-
-
 # functions for aggregate
 def grouped_sum():
     return pysql.sum
